@@ -30,12 +30,13 @@ func (t *TargetFlags) AddTargetFlags(cmd *cobra.Command) {
 func (t *TargetFlags) AddTargetFlagsRules(cmd *cobra.Command) {
 	cmd.Flags().BoolVarP(&t.IsGlobal, "global", "g", false, "Retrive rules that apply globally.")
 	cmd.Flags().StringVarP(&t.MachineID, "machine", "m", "", "Retrieve rules for a single machine. Omit to apply to the current machine.")
-
 }
 
 func (t TargetFlags) GetMachineID() (string, error) {
-	t.getSelfMachineID() // We do some ninja initialization
-
+	_, err := t.getSelfMachineID() // We do some ninja initialization
+	if err != nil {
+		return "", err
+	}
 	if t.IsGlobal {
 		return "", errors.New("do not call GetMachineID when IsGlobal is true")
 	} else if t.MachineID != "" {
